sample demo to show the empty Array issue
Issue detail: Although all yml properties object values are returned as String in Ymal processing, no indexing is available for empty array in the map. Hence it gets picked up like a String property in the code later. Ideally we should get it as indexed value as is the case for emptyElementArray.